Moved streams to separate project and added reactive-streams dependency
