+++ vertx-lang-scala-streams/src/main/scala/io/vertx/lang/scala/stream/builder/StreamBuilder.scala
